backend/requirements.txt
dj-rest-auth[with_social]==4.0.1
django-allauth==0.54.0
django-cors-headers==4.2.0
django==4.2.5
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
pillow==10.0.0
psycopg2-binary==2.9.7

